Delete機能修正、Jestテスト追加

+++ ms1/src/cart/application/domain/service/delete-to-cart.service.ts
@@ -20,6 +20,13 @@ export class DeleteToCartService implements DeleteToCartUseCase {
    * @returns
    */
   public async deleteItem(cmd: DeleteToCartCommand): Promise<void | ThisError> {
+    // 削除対象のCartを取得する
+    const cart = await this.cartRepository.findOne(cmd.cartId);
+    // Cartが見つからない場合はエラーを返す
+    if (!cart) {
+      // return new Exception('カートが見つかりません');
+      return ThisError.invalidArgumentError('削除対象のカートが見つかりません。');
+    }
     // Cartを削除する
     await this.cartRepository.deleteItem({
       cartId: cmd.cartId,
